mpris: Rewrite entirely, replacing pydbus with jeepney
*pydbus* had issues with properly registering, unregistering and filtering
dbus signals, resulting in one listener being created for each player
recognized and never cleaning them up. It is also unmaintained and the
Github repository hasn't seen activity (including responses to issues
or pull requests for over 4 years).

Additionally, there were several situations where opening another player
in the background with the same or a different file (that may or may
not be a recognizable episode) would trip up the tracker entirely,
resulting in missed updates that would cascade to any following episode
of the same show.

These situations are now resolved by:

1. Switching to *jeepney*, a more low-level dbus library, providing full
   access to the dbus signal subscriber & filtering protocol.
   This way we only need a total of two watchers at all times.
2. Properly tracking the full state of every player that fulfills the
   regex criteria.
3. Ignoring any updates while the current player with a valid episode is
   still playing.
4. Ignoring any updates for players whose filename cannot be recognized.

Because the parent Tracker class was not built with the idea of
potentially handling multiple players at the same time, this can still
result in confusions when the primary player is paused and another
valid episode is opened in the background, which will then replace the
tracking status of the previously paused player. However, this can be
recovered if the newly opened player is closed in time so that the
intended primary player has enough time to reach the episode update
threshold afterwards (and will probably occur rarely enough).

The code utilizes asyncio and dataclasses, bumping the required Python
version to 3.7, unless a dataclass backport is added. However, Python
3.7 is the oldest minor version that hasn't reached EOL yet, so this
should be of no concern.

Additionally, dependencies are re-locked due to the switch to jeepney.

Closes #440
Closes #420
Closes #636